+++ theories/proofmode/reduction.v
@@ -6,7 +6,8 @@ From iris.proofmode Require Import base environments.
     do not reduce e.g. before unification happens in [iApply].*)
 Declare Reduction pm_eval := cbv [
   (* base *)
-  base.beq base.Pos_succ base.ascii_beq base.string_beq base.positive_beq base.ident_beq
+  base.negb base.beq
+  base.Pos_succ base.ascii_beq base.string_beq base.positive_beq base.ident_beq
   (* environments *)
   env_lookup env_lookup_delete env_delete env_app env_replace
   env_dom env_intuitionistic env_spatial env_counter env_spatial_is_nil envs_dom
@@ -15,8 +16,8 @@ Declare Reduction pm_eval := cbv [
   envs_clear_spatial envs_clear_intuitionistic envs_incr_counter
   envs_split_go envs_split
   env_to_prop_go env_to_prop env_to_prop_and_go env_to_prop_and
-  (* PM option combinators *)
-  pm_option_bind pm_from_option pm_option_fun
+  (* PM list and option functions *)
+  pm_app pm_option_bind pm_from_option pm_option_fun
 ].
 Ltac pm_eval t :=
   eval pm_eval in t.